size_t->uint64, fix type error

+++ driver/InputFlags.hpp
@@ -111,9 +111,9 @@ public:
     TensorParameters GetValueTensor(const std::string& long_name) const;
     TensorParametersUint64 GetValueTensorUint64(const std::string& long_name) const;
     std::vector<int32_t> GetValueVectorInt(const std::string& long_name) const;
-    std::vector<size_t> GetValueVectorUint64(const std::string& long_name) const;
+    std::vector<uint64_t> GetValueVectorUint64(const std::string& long_name) const;
     std::vector<std::vector<int32_t>> GetValue2dVectorInt(const std::string& long_name) const;
-    std::vector<std::vector<size_t>> GetValue2dVectorUint64(const std::string& long_name) const;
+    std::vector<std::vector<uint64_t>> GetValue2dVectorUint64(const std::string& long_name) const;
     void SetValue(const std::string& long_name, const std::string& new_value);
     void StoreOptionalFlagValue(char short_name, const std::string& input_value);
